+++ MapboxSearch/offline/src/test/java/com/mapbox/search/offline/OfflineSearchResultTest.kt
@@ -1,14 +1,14 @@
+@file:Suppress("DEPRECATION")
+
 package com.mapbox.search.offline
 
 import com.mapbox.geojson.Point
-import com.mapbox.search.base.logger.reinitializeLogImpl
-import com.mapbox.search.base.logger.resetLogImpl
+import com.mapbox.search.base.SdkAssertion
 import com.mapbox.search.base.result.BaseRawResultType
 import com.mapbox.search.base.result.BaseRawSearchResult
 import com.mapbox.search.base.utils.extension.mapToPlatform
 import com.mapbox.search.common.tests.CustomTypeObjectCreatorImpl
 import com.mapbox.search.common.tests.ReflectionObjectsFactory
-import com.mapbox.search.common.tests.TestConstants
 import com.mapbox.search.common.tests.ToStringVerifier
 import com.mapbox.search.common.tests.catchThrowable
 import com.mapbox.search.common.tests.createTestCoreRoutablePoint
@@ -18,13 +18,27 @@ import com.mapbox.search.offline.tests_support.createTestBaseSearchAddress
 import com.mapbox.test.dsl.TestCase
 import io.mockk.every
 import io.mockk.mockk
-import io.mockk.unmockkStatic
+import io.mockk.mockkObject
+import io.mockk.unmockkObject
 import nl.jqno.equalsverifier.EqualsVerifier
-import org.junit.jupiter.api.AfterAll
-import org.junit.jupiter.api.BeforeAll
+import org.junit.jupiter.api.AfterEach
+import org.junit.jupiter.api.BeforeEach
 import org.junit.jupiter.api.TestFactory
 
 internal class OfflineSearchResultTest {
+
+    private val sdkAssertion = mockk<SdkAssertion>(relaxed = true)
+
+    @BeforeEach
+    fun setUp() {
+        mockkObject(SdkAssertion.Companion)
+        every { SdkAssertion.IMPL } returns sdkAssertion
+    }
+
+    @AfterEach
+    fun tearDown() {
+        unmockkObject(SdkAssertion.Companion)
+    }
 
     @TestFactory
     fun `Check OfflineSearchResult equals(), hashCode(), and toString()`() = TestCase {
@@ -58,7 +72,7 @@ internal class OfflineSearchResultTest {
     }
 
     @TestFactory
-    fun `Check OfflineSearchResult instantiation`() = TestCase {
+    fun `Check OfflineSearchResult instantiation with a null coordinate`() = TestCase {
         Given("OfflineSearchResult constructor") {
             When("OfflineSearchResult instantiated with a null coordinate") {
                 val e = catchThrowable<IllegalStateException> {
@@ -68,8 +82,34 @@ internal class OfflineSearchResultTest {
                 Then(
                     "Exception should be thrown",
                     true,
-                    IllegalStateException("Server search result must have a coordinate").equalsTo(e)
+                    IllegalStateException("Offline search result must have a coordinate").equalsTo(e)
                 )
+            }
+        }
+    }
+
+    @TestFactory
+    fun `Check OfflineSearchResult instantiation with unsupported type`() = TestCase {
+        Given("OfflineSearchResult constructor") {
+            When("OfflineSearchResult instantiated with ${BaseRawResultType.LOCALITY} type") {
+                val baseResult = TEST_BASE_RAW_RESULT_1.copy(types = listOf(BaseRawResultType.LOCALITY))
+                val offlineSearchResult = OfflineSearchResult(baseResult)
+
+                Then(
+                    "newType should be ${NewOfflineSearchResultType.FALLBACK_TYPE}",
+                    NewOfflineSearchResultType.FALLBACK_TYPE,
+                    offlineSearchResult.newType,
+                )
+
+                Then(
+                    "type should be ${OfflineSearchResultType.DEFAULT}",
+                    OfflineSearchResultType.DEFAULT,
+                    offlineSearchResult.type,
+                )
+
+                Verify("assertDebug() called", atLeast = 1) {
+                    sdkAssertion.assertDebug(any(), any())
+                }
             }
         }
     }
@@ -123,7 +163,9 @@ internal class OfflineSearchResultTest {
                     base.routablePoints
                 }
 
-                Then("type should be $TEST_TYPE", TEST_TYPE, searchResult.type)
+                val oldType = NewOfflineSearchResultType.toOldResultType(searchResult.newType)
+                Then("type should be $oldType", oldType, searchResult.type)
+                Then("newType should be $TEST_TYPE", TEST_TYPE, searchResult.newType)
                 VerifyOnce("base.types called") {
                     base.types
                 }
@@ -147,7 +189,7 @@ internal class OfflineSearchResultTest {
         val TEST_CORE_ROUTABLE_POINT = createTestCoreRoutablePoint()
         val TEST_ROUTABLE_POINT = TEST_CORE_ROUTABLE_POINT.mapToPlatform()
         val TEST_BASE_TYPE = BaseRawResultType.ADDRESS
-        val TEST_TYPE = TEST_BASE_TYPE.tryMapToOfflineSdkType()
+        val TEST_TYPE = NewOfflineSearchResultType.createFromRawResultType(TEST_BASE_TYPE)
         const val TEST_DISTANCE_METERS = 123.456
 
         val TEST_BASE_RAW_RESULT_1 = createTestBaseRawSearchResult(
@@ -163,20 +205,5 @@ internal class OfflineSearchResultTest {
             center = Point.fromLngLat(30.0, 50.0),
             types = listOf(BaseRawResultType.STREET)
         )
-
-        @Suppress("JVM_STATIC_IN_PRIVATE_COMPANION")
-        @BeforeAll
-        @JvmStatic
-        fun setUpAll() {
-            resetLogImpl()
-        }
-
-        @Suppress("JVM_STATIC_IN_PRIVATE_COMPANION")
-        @AfterAll
-        @JvmStatic
-        fun tearDownAll() {
-            reinitializeLogImpl()
-            unmockkStatic(TestConstants.ASSERTIONS_KT_CLASS_NAME)
-        }
     }
 }